update codebase with variable name and create new file for animation module
