+++ Akshay Sakare/app/src/main/java/com/apsakare/moviesapp/MainActivity.kt
@@ -1,5 +1,6 @@
 package com.apsakare.moviesapp
 
+import ComposeLottieScreen
 import android.annotation.SuppressLint
 import android.content.Context
 import android.net.ConnectivityManager
@@ -62,171 +63,11 @@ class MainActivity : ComponentActivity() {
                 Surface(
                     modifier = Modifier.fillMaxSize(),
                     color = app_black
-                ) { //Surface(modifier = Modifier.fillMaxSize(), color = colorScheme.background) {
-                    // Lottie animation screen.
+                ) { // Lottie animation screen.
                     ComposeLottieScreen()
-                    //}
                 }
             }
         }
     }
 }
 
-@Composable
-fun ComposeLottieScreen() {
-
-    Surface(
-        modifier = Modifier.fillMaxSize(),
-
-        ) {
-        val composition by rememberLottieComposition(
-            spec = LottieCompositionSpec.RawRes(R.raw.cinema)
-        )
-        var isPlaying by remember {
-            mutableStateOf(true)
-        }
-        val progress by animateLottieCompositionAsState(
-            composition = composition,
-            isPlaying = isPlaying
-        )
-        val sContext = LocalContext.current
-        LaunchedEffect(key1 = progress) {
-
-            if (progress == 0f) {
-                isPlaying = true
-
-            }
-            if (progress == 1f) {
-                isPlaying = false
-            }
-
-
-        }
-        if (!isPlaying) {
-            Surface(modifier = Modifier.fillMaxSize(), color = colorScheme.background) {
-                if (isInternetAvailable(LocalContext.current)) {
-                    ShowSplash()
-                } else {
-                    val openDialog = remember { mutableStateOf(false) }
-                    openDialog.value = true
-                    AlertDialogNetwork(openDialog)
-                }
-            }
-        }
-
-        LottieAnimation(
-            composition = composition,
-            modifier = Modifier
-                .size(1500.dp)
-                .clickable {
-                    isPlaying = true
-                },
-
-            progress = {
-                progress
-            }
-        )
-    }
-}
-
-@Composable
-fun ShowSplash() {
-    Box(modifier = Modifier.fillMaxSize()) {
-        Image(
-            painter = painterResource(id = R.drawable.icon),
-            contentDescription = "image",
-            alignment = Alignment.Center,
-            modifier = Modifier.matchParentSize()
-        )
-        Column(
-            modifier = Modifier.fillMaxSize(),
-            verticalArrangement = Arrangement.SpaceEvenly,
-            horizontalAlignment = CenterHorizontally,
-
-            ) {
-            val sContext = LocalContext.current
-            Text(
-                text = sContext.getString(R.string.welcome),
-                color = White,
-                fontWeight = FontWeight.Bold,
-                fontSize = 35.sp
-            )
-            Spacer(modifier = Modifier.width(10.dp))
-            Button(
-                onClick = {
-                    Toast.makeText(
-                        sContext,
-                        sContext.getText(R.string.splash_btn),
-                        Toast.LENGTH_LONG
-                    ).show()
-                },
-                modifier = Modifier
-                    .padding(16.dp)
-                    .align(CenterHorizontally),
-
-                colors = ButtonDefaults.buttonColors(golden)
-            ) {
-                Text(
-                    modifier = Modifier.padding(10.dp),
-                    text = sContext.getString(R.string.splash_btn), color = Black,
-                    fontSize = 15.sp
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun AlertDialogNetwork(openDialog: MutableState<Boolean>) {
-    val sContext = LocalContext.current
-    if (openDialog.value) {
-        AlertDialog(
-            onDismissRequest = {
-                openDialog.value = false
-            },
-            title = {
-                Text(text = sContext.getString(R.string.no_internet))
-            },
-            text = {
-                Text(sContext.getString(R.string.check_internet))
-            },
-            confirmButton = {
-                Button(
-                    onClick = {
-                        openDialog.value = false
-                    }) {
-                    Text(sContext.getString(R.string.okay))
-                }
-            },
-
-            )
-    }
-}
-
-private fun isInternetAvailable(context: Context): Boolean {
-    var result = false
-    val connectivityManager =
-        context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-        val networkCapabilities = connectivityManager.activeNetwork ?: return false
-        val actNw = connectivityManager.getNetworkCapabilities(networkCapabilities) ?: return false
-        result = when {
-            actNw.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> true
-            actNw.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> true
-            actNw.hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET) -> true
-            else -> false
-        }
-    } else {
-        connectivityManager.run {
-            connectivityManager.activeNetworkInfo?.run {
-                result = when (type) {
-                    ConnectivityManager.TYPE_WIFI -> true
-                    ConnectivityManager.TYPE_MOBILE -> true
-                    ConnectivityManager.TYPE_ETHERNET -> true
-                    else -> false
-                }
-            }
-        }
-    }
-    return result
-}
